+++ R/content.R
@@ -93,11 +93,25 @@ Content <- R6::R6Class(
     get_dashboard_url = function(pane = "") {
       dashboard_url_chr(self$connect$server, self$content$guid, pane = pane)
     },
-    #' @description Return the jobs for this content.
+    #' @description Return the jobs for this content
     jobs = function() {
-      warn_experimental("jobs")
-      url <- unversioned_url("applications", self$get_content()$guid, "jobs")
-      res <- self$get_connect()$GET(url)
+      res <- self$connect$GET(v1_url("content", self$content$guid, "jobs"), parser = NULL)
+      use_unversioned <- endpoint_does_not_exist(res)
+      if (use_unversioned) {
+        res <- self$connect$GET(unversioned_url("applications", self$content$guid, "jobs"), parser = NULL)
+      }
+      self$connect$raise_error(res)
+      parsed <- httr::content(res, as = "parsed")
+      if (use_unversioned) {
+        # The unversioned endpoint does not contain a `status` field. Its field
+        # `finalized` is `FALSE` corresponds to active jobs. The `finalized`
+        # field is dropped during parsing.
+        parsed <- purrr::modify_if(parsed, ~ isFALSE(.x$finalized), function(x) {
+          x$status <- 0
+          x
+        })
+      }
+      parsed
     },
     #' @description Return a single job for this content.
     #' @param key The job key.
@@ -111,6 +125,17 @@ Content <- R6::R6Class(
         list(res),
         ~ purrr::list_modify(.x, app_guid = content_guid)
       )[[1]]
+    },
+    #' @description Terminate a single job for this content item.
+    #' @param key The job key.
+    register_job_kill_order = function(key) {
+      con <- self$connect
+      url <- v1_url("content", self$content$guid, "jobs", key)
+      res <- self$connect$DELETE(url)
+      if (endpoint_does_not_exist(res)) {
+        con$raise_error(res)
+      }
+      res
     },
     #' @description Return the variants for this content.
     variants = function() {
@@ -588,26 +613,73 @@ content_ensure <- function(
 
 #' Get Jobs
 #'
-#' `r lifecycle::badge('experimental')` Retrieve details about jobs associated with a `content_item`.
-#' "Jobs" in Posit Connect are content executions
+#' Retrieve details about server processes associated with a `content_item`,
+#' such as a FastAPI app or a Quarto render.
+#'
+#' Note that Connect versions below 2022.10.0 use a legacy endpoint, and will
+#' not return the complete set of information provided by newer versions.
+#'
+#' @param content A Content object, as returned by `content_item()`
+#'
+#' @return A data frame with a row for each job, with the following columns:
+#'
+#' - `id`: The job identifier.
+#' - `ppid`: The job's parent process identifier (see Note 1).
+#' - `pid`: The job's process identifier.
+#' - `key`: The job's unique key identifier.
+#' - `remote_id`: The job's identifier for off-host execution configurations
+#' (see Note 1).
+#' - `app_id`: The job's parent content identifier
+#' - `variant_id`: The identifier of the variant owning this job.
+#' - `bundle_id`: The identifier of a content bundle linked to this job.
+#' - `start_time`: The timestamp (RFC3339) indicating when this job started.
+#' - `end_time`: The timestamp (RFC3339) indicating when this job finished.
+#' - `last_heartbeat_time`: The timestamp (RFC3339) indicating the last time
+#' this job was observed to be running (see Note 1).
+#' - `queued_time`: The timestamp (RFC3339) indicating when this job was added
+#' to the queue to be processed. Only scheduled reports will present a value
+#' for this field (see Note 1).
+#' - `queue_name`: The name of the queue which processes the job. Only
+#' scheduled reports will present a value for this field (see Note 1).
+#' - `tag`: A tag to identify the nature of the job.
+#' - `exit_code`: The job's exit code. Present only when job is finished.
+#' - `status`: The current status of the job. On Connect 2022.10.0 and newer,
+#' one of Active: 0, Finished: 1, Finalized: 2; on earlier versions, Active:
+#' 0, otherwise `NA`.
+#' - `hostname`: The name of the node which processes the job.
+#' - `cluster`: The location where this content runs. Content running on the
+#' same server as Connect will have either a null value or the string Local.
+#' Gives the name of the cluster when run external to the Connect host
+#' (see Note 1).
+#' - `image`: The location where this content runs. Content running on
+#' the same server as Connect will have either a null value or the string
+#' Local. References the name of the target image when content runs in
+#' a clustered environment such as Kubernetes (see Note 1).
+#' - `run_as`: The UNIX user that executed this job.
+#'
+#' @note
+#' 1. On Connect instances earlier than 2022.10.0, these columns will contain `NA` values.
+#'
+#' @family job functions
+#' @family content functions
+#' @export
+get_jobs <- function(content) {
+  validate_R6_class(content, "Content")
+
+  jobs <- content$jobs()
+  parse_connectapi_typed(jobs, connectapi_ptypes$jobs, strict = TRUE)
+}
+
+# TODO: Need to test `logged_error` on a real error
+#'
+#' Retrieve details about a server process
+#' associated with a `content_item`, such as a FastAPI app or a Quarto render.
 #'
 #' @param content A Content object, as returned by `content_item()`
 #' @param key The key for a job
 #'
-#' @rdname jobs
+#' @family job functions
 #' @family content functions
-#' @export
-get_jobs <- function(content) {
-  warn_experimental("get_jobs")
-  scoped_experimental_silence()
-  validate_R6_class(content, "Content")
-
-  jobs <- content$jobs()
-  parse_connectapi_typed(jobs, connectapi_ptypes$jobs)
-}
-
-# TODO: Need to test `logged_error` on a real error
-#' @rdname jobs
 #' @export
 get_job <- function(content, key) {
   warn_experimental("get_job")
@@ -621,6 +693,58 @@ get_job <- function(content, key) {
   # a bit of an abuse
   # since stdout / stderr / logged_error are here now...
   parse_connectapi_typed(list(job), connectapi_ptypes$job)
+}
+
+#' Terminate Jobs
+#'
+#' Register a job kill order for one or more jobs associated with a content
+#' item. Requires Connect 2022.10.0 or newer.
+#'
+#' @param content A Content object, as returned by `content_item()`
+#' @param keys Optional. One or more job keys, which can be obtained using
+#' `get_jobs(content)`. If no keys are provided, will terminate all active
+#' jobs for the provided content item.
+
+#' @return A data frame with the status of each termination request.
+#'
+#' - `app_id`: The content item's identifier.
+#' - `app_guid`: The content item's GUID.
+#' - `job_key`: The job key.
+#' - `job_id`: The job's identifier.
+#' - `result`: The result string returned by Connect.
+#' - `code`: An error code, `NA` if the request was successful.
+#' - `error`: An error message, `NA` if the result was successful.
+#'
+#' Note that `app_id`, `app_guid`, `job_id`, and `result` are `NA` if the
+#' request returns an error.
+#'
+#' @family job functions
+#' @family content functions
+#' @export
+terminate_jobs <- function(content, keys = NULL) {
+  validate_R6_class(content, "Content")
+
+  if (is.null(keys)) {
+    all_jobs <- get_jobs(content)
+    keys <- all_jobs[all_jobs$status == 0, ]$key
+    if (length(keys) == 0) {
+      message("No active jobs found.")
+      return(vctrs::vec_ptype(connectapi_ptypes$job_termination))
+    }
+  }
+
+  res <- purrr::map(keys, content$register_job_kill_order)
+  res_content <- purrr::map(res, httr::content)
+  res_df <- tibble::tibble(
+    parse_connectapi_typed(
+      res_content,
+      connectapi_ptypes$job_termination,
+      strict = TRUE
+    )
+  )
+  # Errors will not have the job_key.
+  res_df$job_key <- keys
+  res_df
 }
 
 #' Set RunAs User